feat: rename Student components to Etudiant for consistency and update related imports
