+++ client/src/components/Etudiant/EtudiantDetails.tsx
